Continue 64-bit port: fixes across engine, gameplay, and build
- Fix x64 portability bugs (pointer-array to int* truncation,
  uninitialized member pointers) across Lawn/, SexyAppFramework/,
  Sexy.TodLib/, and PakLib
- Gameplay/widget updates: Board, Plant, Zombie, Projectile, Coin,
  LawnMower, Challenge, CutScene, SeedPacket, and UI dialogs
  (Almanac, AwardScreen, ChallengeScreen, ContinueDialog, SeedChooser)
- Update build system (CMakeLists.txt, build_linux.sh) and add
  .clang-format
- Reorganize scratch files into docs/reverse-engineering/ and tools/

+++ Sexy.TodLib/DataArray.h
@@ -116,7 +116,7 @@ public:
 			aItem++;
 
 		DataArray<T>::DataArrayItem* aLast = &mBlock[mMaxUsedCount];
-		while ((unsigned int)aItem < (unsigned int)aLast)
+		while (aItem < aLast)
 		{
 			if (aItem->mID & DATA_ARRAY_KEY_MASK)
 			{
